[proliferate] split workspace shell components
