+++ desktop/src/hooks/workspaces/ui/use-right-panel-state-updater.ts
@@ -1,0 +1,41 @@
+import {
+  useCallback,
+  type Dispatch,
+  type SetStateAction,
+} from "react";
+import {
+  reconcileRightPanelWorkspaceState,
+  type RightPanelWorkspaceState,
+} from "@/lib/domain/workspaces/shell/right-panel";
+import { rightPanelStateEqual } from "@/lib/domain/workspaces/shell/right-panel-view";
+
+export function useRightPanelStateUpdater({
+  isCloudWorkspaceSelected,
+  onStateChange,
+}: {
+  isCloudWorkspaceSelected: boolean;
+  onStateChange: Dispatch<SetStateAction<RightPanelWorkspaceState>>;
+}) {
+  return useCallback(
+    (value: SetStateAction<RightPanelWorkspaceState>) => {
+      onStateChange((previous) => {
+        const current = reconcileRightPanelWorkspaceState(previous, {
+          isCloudWorkspaceSelected,
+        });
+        const next = typeof value === "function"
+          ? (value as (previousValue: RightPanelWorkspaceState) => RightPanelWorkspaceState)(
+              current,
+            )
+          : value;
+        const reconciledNext = reconcileRightPanelWorkspaceState(next, {
+          isCloudWorkspaceSelected,
+        });
+        if (!rightPanelStateEqual(current, reconciledNext)) {
+          return reconciledNext;
+        }
+        return rightPanelStateEqual(previous, current) ? previous : current;
+      });
+    },
+    [isCloudWorkspaceSelected, onStateChange],
+  );
+}
